Add ruby to the list of homebrew dependencies for Mac
The guide seems to assume that the reader already has Ruby installed through homebrew. If that's not the case, `gem update --system` will use Mac OS's system ruby, which exits with a file permissions error (and may lead the reader to fix that with sudo/chown, ending up with a different setup than is described in the guide).